6_Reorganized code in 5 components, added connections buttons
