+++ src/app/header/back-button/back-button.component.css
@@ -1,0 +1,25 @@
+
+.header__text__unselectable {
+	-webkit-touch-callout: none; /* iOS Safari */
+	-webkit-user-select: none; /* Safari */
+	-khtml-user-select: none; /* Konqueror HTML */
+	-moz-user-select: none; /* Firefox */
+	-ms-user-select: none; /* Internet Explorer/Edge */
+	user-select: none; /* Non-prefixed version, currently supported by Chrome and Opera */
+}
+
+.menu__button {
+	min-width: 64px;
+	min-height: 64px;
+}
+
+.menu__button:hover {
+	cursor: pointer;
+}
+
+.menu__buttonIcon {
+	max-width: 14px;
+	max-height: 14px;
+	font-size: 18px;
+	margin: 25px;
+}
