+++ src/app/header/back-button/back-button.component.html
@@ -1,0 +1,7 @@
+<div
+	 class="menu__button header__text__unselectable"
+	 (click)="navigateToHome()">
+	<mat-icon class="menu__buttonIcon">
+		arrow_back
+	</mat-icon>
+</div>
