+++ src/app/header/connection-buttons/connection-buttons.component.css
@@ -1,0 +1,30 @@
+
+.header__text__unselectable {
+	-webkit-touch-callout: none; /* iOS Safari */
+	-webkit-user-select: none; /* Safari */
+	-khtml-user-select: none; /* Konqueror HTML */
+	-moz-user-select: none; /* Firefox */
+	-ms-user-select: none; /* Internet Explorer/Edge */
+	user-select: none; /* Non-prefixed version, currently supported by Chrome and Opera */
+}
+.header__connection{
+	position: absolute;
+	right: 0px;
+	top: -10px;
+	display: flex;
+	align-items: center;
+}
+
+.header__connectionText {
+	font-size: 14px;
+	font-weight: 500;
+	font-style: normal;
+	font-stretch: normal;
+	line-height: normal;
+	letter-spacing: 0.5px;
+	cursor: pointer;
+}
+
+.header__connectionText__last{
+	margin-right: 42px;
+}
