+++ src/app/header/connection-buttons/connection-buttons.component.html
@@ -1,0 +1,13 @@
+<div class="header__connection">
+	<a (click)="navigateToLogin()"
+	   class="header__connectionText">
+		Se connecter
+	</a>
+	<a class="header__connectionText">
+		&nbsp;|&nbsp;
+	</a>
+	<a (click)="navigateToSignIn()"
+	   class="header__connectionText header__connectionText__last">
+		S'inscrire
+	</a>
+</div>
