+++ src/app/header/side-menu/side-menu.component.css
@@ -1,23 +1,11 @@
-.header {
-	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-	padding: 0px;
-	color: white;
-	background-color: #4a4a4a;
-	max-height: 64px;
-	font-family: Roboto;
-	font-size: 36px;
-	font-weight: normal;
-	font-style: normal;
-	font-stretch: normal;
-	line-height: 0.89;
-}
 
-
-
-.header__menu {
-	position: relative;
+.header__text__unselectable {
+	-webkit-touch-callout: none; /* iOS Safari */
+	-webkit-user-select: none; /* Safari */
+	-khtml-user-select: none; /* Konqueror HTML */
+	-moz-user-select: none; /* Firefox */
+	-ms-user-select: none; /* Internet Explorer/Edge */
+	user-select: none; /* Non-prefixed version, currently supported by Chrome and Opera */
 }
 
 .menu__button {
@@ -71,10 +59,3 @@
 .menu__choices .menu__choice:hover {
 	color: #f5a623;
 }
-
-
-.header__options {
-	position: relative;
-}
-
-
