+++ src/app/header/side-menu/side-menu.component.html
@@ -1,0 +1,35 @@
+
+<div (click)="handleMenuButtonClick()">
+	<div class="menu__button header__text__unselectable">
+		<mat-icon class="menu__buttonIcon">
+			{{sideMenuIcon}}
+		</mat-icon>
+	</div>
+
+	<div *ngIf="openSideMenu"
+		 class="menu__choices">
+		<div class="menu__choice"
+			 (click)="navigateToIndex()">
+			<mat-icon class="menu__choiceIcon">
+				move_to_inbox
+			</mat-icon>
+			Accueil
+		</div>
+
+		<div class="menu__choice"
+			 (click)="navigateToEvents()">
+			<mat-icon class="menu__choiceIcon">
+				send
+			</mat-icon>
+			Évènements
+		</div>
+
+		<div class="menu__choice"
+			 (click)="navigateToContact()">
+			<mat-icon class="menu__choiceIcon">
+				group
+			</mat-icon>
+			Contact
+		</div>
+	</div>
+</div>
