--- src/app/options/options.component.html
@@ -1,3 +1,0 @@
-<p>
-  options works!
-</p>
