--- src/app/side-menu/side-menu.component.html
@@ -1,3 +1,0 @@
-<p>
-  side-menu works!
-</p>
