--- src/app/side-menu/side-menu.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-side-menu',
-  templateUrl: './side-menu.component.html',
-  styleUrls: ['./side-menu.component.css']
-})
-export class SideMenuComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
